FIX: issue #17 (#19)
The module tries to read the label of the resource template even when no template has been selected by the user. That causes error's when creating Items w/o selecting a template (e.g. when adding an item through the Collecting module or just leaving the dropdown as it is when creating a new Item. This fix sets the label to null when no template has been selected.